requirements.txt
environs==9.5.0
google-cloud-api-keys==0.5.1
google-cloud-dialogflow==2.20.0
langid==1.1.6
python-telegram-bot==13.15
vk-api==11.9.9
